Fix hw timers / sw timers in mbed (Now working!)

+++ src/ports/lpc17xx/common.h
@@ -166,9 +166,9 @@ extern "C" {
 
 // Scheduler structure
     struct _scheduler_t {
-        uint8_t timer_id;
-        uint8_t mr_id;
-        uint8_t sw_mr_id;
+        uint32_t timer_id;
+        uint32_t mr_id;
+        uint32_t sw_mr_id;
     };
 
     struct _i2c_t {
